[eve.trader](feat/acceptance): accept the pipeline by rebuilding features from raw
Приёмка конвейера наблюдения: телеметрия по контракту OTel, отчёт о расходе
ресурсов, десять команд оператора и приёмочный сценарий `accept` — перестроить
признаки за интервал из сырья через реплей и сверить с записанными.

Прогон на архиве EVE Ref (15 сентября 2026, два хаба, десять снимков) сошёлся:
1 176 760 признаков на шестнадцати наблюдениях, ноль расхождений, интервал
содержит и смену версии статических данных, и досинхронизацию задним числом.

По дороге найдены и исправлены две ошибки:

- Восстановление сравнивало две шкалы времени: якорем брался чекпойнт по времени
  наблюдения, а события — по времени события. Перестановка, чей `issued` попадал
  внутрь интервала сбора предыдущего снимка, выпадала из восстановления навсегда.
- Цепочка наблюдений считалась по региону, а не по паре «регион и источник»:
  суточная история объявляла разрыв там, где данные полны.

Остаточное расхождение разобрано измерением на десяти настоящих снимках The Forge
и оказалось не дефектом: последнее наблюдение цепочки невоспроизводимо по
построению — окно подтверждения исчезновений не закрылось. Норма
`market-observation/book-features` дополнена этой оговоркой, разметка событий не
тронута.

Замеры против бюджета сняты и записаны; пиковая память расходится в 14–16 раз и
ждёт решения владельца.

+++ tests/integration/EveTrader.Application.Integration.Workers/ScheduledWorkerHostingShould.cs
@@ -18,7 +18,7 @@ public sealed class ScheduledWorkerHostingShould
     [Fact]
     public async Task RunCyclesUnderHostAndEmitCountersToMeter()
     {
-        var measurements = new ConcurrentBag<(string Instrument, long Value)>();
+        var measurements = new ConcurrentBag<(string Instrument, double Value)>();
 
         using var listener = new MeterListener
         {
@@ -31,7 +31,9 @@ public sealed class ScheduledWorkerHostingShould
             },
         };
 
-        listener.SetMeasurementEventCallback<long>(
+        // Инструменты объявлены в double: длительности и доли целыми не бывают, а
+        // один тип замера избавляет экспорт от разбора, каким он приехал.
+        listener.SetMeasurementEventCallback<double>(
             (instrument, value, _, _) => measurements.Add((instrument.Name, value)));
         listener.Start();
 
@@ -54,7 +56,7 @@ public sealed class ScheduledWorkerHostingShould
 
         measurements.Where(measurement => measurement.Instrument == "worker.Ticking.ticked")
             .Sum(measurement => measurement.Value)
-            .ShouldBeGreaterThanOrEqualTo(3);
+            .ShouldBeGreaterThanOrEqualTo(3d);
 
         measurements.ShouldContain(measurement => measurement.Instrument == "worker.Ticking.outcome.succeeded");
         measurements.ShouldNotContain(measurement => measurement.Instrument == "worker.Ticking.outcome.failed");
